[Core] Move email managers to mailer directory
