+++ src/Sylius/Bundle/CoreBundle/Mailer/ShipmentEmailManager.php
@@ -11,28 +11,32 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\CoreBundle\EmailManager;
+namespace Sylius\Bundle\CoreBundle\Mailer;
 
-use Sylius\Bundle\CoreBundle\Mailer\Emails;
 use Sylius\Component\Core\Model\OrderInterface;
+use Sylius\Component\Core\Model\ShipmentInterface;
 use Sylius\Component\Mailer\Sender\SenderInterface;
 use Webmozart\Assert\Assert;
 
-final class OrderEmailManager implements OrderEmailManagerInterface
+final class ShipmentEmailManager implements ShipmentEmailManagerInterface
 {
     public function __construct(private SenderInterface $emailSender)
     {
     }
 
-    public function resendConfirmationEmail(OrderInterface $order): void
+    public function resendConfirmationEmail(ShipmentInterface $shipment): void
     {
+        /** @var OrderInterface $order */
+        $order = $shipment->getOrder();
+        Assert::notNull($order);
         $email = $order->getCustomer()->getEmail();
         Assert::notNull($email);
 
         $this->emailSender->send(
-            Emails::ORDER_CONFIRMATION_RESENT,
+            Emails::SHIPMENT_CONFIRMATION_RESENT,
             [$email],
             [
+                'shipment' => $shipment,
                 'order' => $order,
                 'channel' => $order->getChannel(),
                 'localeCode' => $order->getLocaleCode(),
